fixes bug and moves helper objects
